# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API Sprint 5",
        "description": "Documentación de la API",
        "contact": {
            "email": "ivonnecp@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Actividades"
                ],
                "summary": "Obtener todas las actividades",
                "operationId": "69c4fdfcb7814c075663f2087e01ea02",
                "responses": {
                    "200": {
                        "description": "Lista de actividades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activities": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Actividades"
                ],
                "summary": "Crear una nueva actividad",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actividad creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activity": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/activities/{activity}": {
            "get": {
                "tags": [
                    "Actividades"
                ],
                "summary": "Obtener una actividad específica",
                "operationId": "51fff196ee83942690de8b440e2ddfbe",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "ID de la actividad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la actividad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activity": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Actividades"
                ],
                "summary": "Actualizar una actividad",
                "operationId": "bda87636e0bb8d2290b9a2280abcdbd5",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "ID de la actividad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activity": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actividades"
                ],
                "summary": "Eliminar una actividad",
                "operationId": "e25deaa842b7124f291bd5245fcab496",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "ID de la actividad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}